Create readme.md
The communication part for the transformation bot is completed.
the problem we face while communication via bluetooth (range and latency) is resolved with esp32 websocket where data are sent over wifi almost instantly. also the webpage is ready and is sending data as expected. the code for motor control is also ready to test.
the electronics and communication part is done for now till the hardware arrives and for testing